exynos: gralloc4: Add support for contiguous planes
The current gralloc implementation assumes that alignment operations
need to be performed per-plane in multiplanar formats as they are
in separate buffers.

When this isn't the case, and the chroma planes follow immediately
after the luma plane, this can result in miscalculated stride values
for some buffer dimensions resulting in visual corruption.

Bug: 150319746
Test: build, boot
Change-Id: I5fa274dc810ae16af6c7bf6873cab1408a7fc42a
Signed-off-by: Majaahh <maja.blaszczak@proton.me>